pyproject.toml
[tool.poetry]
name = "py-push-arrivals-and-departures"
version = "1.0.0"
description = ""
authors = ["Alyosha <alyosha@30emeciel.fr>"]

[tool.poetry.dependencies]
python = "^3.9"
python-box = "^5.3.0"
core = { git = "https://github.com/30emeciel/py-core.git", branch="4"}


[tool.poetry.dev-dependencies]
python-dotenv = "^0.17.0"
pytest = "^6.2.4"
pytest-mockito = "^0.0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
